Documents/data_science/Streamlit/app_tele/requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.0.0
xlsxwriter>=3.0.0
openpyxl>=3.0.0
